refactor: split engine.rs into directory module to fix readability ratchet
engine.rs grew to 1101 lines (baseline max 987). Split into:
- engine/mod.rs (568 lines) — core AuthzEngine implementation
- engine/tests.rs (532 lines) — all 50 unit tests
